refactor(tui): split chat model.go by concern, matching the debug layout
model.go had grown to 1575 lines holding the model, message types, styles,
the update loop, the agent/streaming loop, session lifecycle, views and
formatting helpers. Split it into model.go, types.go, styles.go, update.go,
stream.go, sessions.go, view.go and helpers.go — the same decomposition the
debug package already uses. Code is moved verbatim; no behavior change.
